+++ docs/public-docs-package/index.js
@@ -5,6 +5,7 @@ module.exports = new Package('angular-v2-public-docs', [basePackage])
 
 .config(function(readTypeScriptModules) {
   readTypeScriptModules.sourceFiles = [
+    'angular2/lifecycle_hooks.ts',
     'angular2/metadata.ts',
     'angular2/change_detection.ts',
     'angular2/core.ts',
