+++ modules/angular2/angular2.ts
@@ -15,3 +15,4 @@ export * from './forms';
 export * from './render';
 export * from './profile';
 export {bootstrap} from 'angular2/src/core/application';
+export * from './lifecycle_hooks';
