+++ modules/angular2/angular2_exports.ts
@@ -7,3 +7,4 @@ export * from './forms';
 export * from './render';
 export * from './profile';
 export {bootstrap} from 'angular2/src/core/application';
+export * from './lifecycle_hooks';
