+++ modules/angular2/lifecycle_hooks.ts
@@ -1,0 +1,17 @@
+/**
+ * @module
+ * @description
+ * Defines interfaces to be implemented by directives when they need to hook into the change
+ * detection mechanism.
+ */
+
+export {
+  AfterContentInit,
+  AfterContentChecked,
+  AfterViewInit,
+  AfterViewChecked,
+  OnChanges,
+  OnDestroy,
+  OnInit,
+  DoCheck
+} from './src/core/compiler/interfaces';
