+++ modules/angular2/metadata.dart
@@ -1,4 +1,3 @@
 library angular2.metadata;
 
 export 'package:angular2/src/core/metadata.dart';
-export 'package:angular2/src/core/compiler/interfaces.dart';
