+++ modules/angular2/metadata.ts
@@ -3,10 +3,9 @@
  * @description
  *
  * Annotations provide the additional information that Angular requires in order to run your
- * application. This module
- * contains {@link ComponentMetadata}, {@link DirectiveMetadata}, and {@link ViewMetadata}
- * annotations, as well as
- * the {@link Host} annotation that is used by Angular to resolve dependencies.
+ * application. This module contains {@link ComponentMetadata}, {@link DirectiveMetadata}, and
+ * {@link ViewMetadata} annotations, as well as the {@link Host} annotation that is used by Angular
+ * to resolve dependencies.
  *
  */
 
@@ -48,17 +47,5 @@ export {
   HostListenerMetadata
 } from './src/core/metadata';
 
-export {
-  // todo(vbe): LifecycleHook should not be exposed (fails test.typings)
-  LifecycleHook,
-  AfterContentInit,
-  AfterContentChecked,
-  AfterViewInit,
-  AfterViewChecked,
-  OnChanges,
-  OnDestroy,
-  OnInit,
-  DoCheck
-} from './src/core/compiler/interfaces';
 
 export {Class, ClassDefinition, TypeDecorator} from './src/core/util/decorators';
