+++ modules/angular2/src/core/directives/ng_style.ts
@@ -1,4 +1,5 @@
-import {Directive, DoCheck} from 'angular2/metadata';
+import {Directive} from 'angular2/metadata';
+import {DoCheck} from 'angular2/lifecycle_hooks';
 import {ElementRef} from 'angular2/core';
 import {KeyValueDiffer, KeyValueDiffers} from 'angular2/change_detection';
 import {isPresent, isBlank, print} from 'angular2/src/core/facade/lang';
