+++ modules/angular2/src/forms/directives/ng_control_group.ts
@@ -1,4 +1,5 @@
-import {Directive, OnInit, OnDestroy} from 'angular2/metadata';
+import {Directive} from 'angular2/metadata';
+import {OnInit, OnDestroy} from 'angular2/lifecycle_hooks';
 import {Inject, Host, SkipSelf, forwardRef, Binding} from 'angular2/di';
 import {ListWrapper} from 'angular2/src/core/facade/collection';
 import {CONST_EXPR} from 'angular2/src/core/facade/lang';
