+++ modules/angular2/src/forms/directives/ng_model.ts
@@ -1,8 +1,9 @@
 import {CONST_EXPR} from 'angular2/src/core/facade/lang';
 import {EventEmitter, ObservableWrapper} from 'angular2/src/core/facade/async';
 
-import {Query, Directive, OnChanges} from 'angular2/metadata';
+import {Query, Directive} from 'angular2/metadata';
 import {forwardRef, Binding, Inject, Optional} from 'angular2/di';
+import {OnChanges} from 'angular2/lifecycle_hooks';
 
 import {NgControl} from './ng_control';
 import {Control} from '../model';
