+++ modules/angular2/test/core/compiler/dynamic_component_loader_spec.ts
@@ -20,7 +20,8 @@ import {
 
 import {Injector} from 'angular2/di';
 import {NgIf} from 'angular2/directives';
-import {Component, View, ViewMetadata, OnDestroy} from 'angular2/metadata';
+import {Component, View, ViewMetadata} from 'angular2/metadata';
+import {OnDestroy} from 'angular2/lifecycle_hooks';
 import {DynamicComponentLoader} from 'angular2/src/core/compiler/dynamic_component_loader';
 import {ElementRef} from 'angular2/src/core/compiler/element_ref';
 import {DOCUMENT} from 'angular2/src/core/render/render';
